feature

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -1,6 +1,7 @@
 import { QuestionsRepository } from '../repositories/questions-repository'
 
 interface DeleteQuestionUseCaseRequest {
+  authorId: string
   questionId: string
 }
 
@@ -10,11 +11,16 @@ export class DeleteQuestionUseCase {
   constructor(private questionsRepository: QuestionsRepository) {}
   async execute({
     questionId,
+    authorId,
   }: DeleteQuestionUseCaseRequest): Promise<DeleteQuestionUseCaseResponse> {
     const question = await this.questionsRepository.findById(questionId)
 
     if (!question) {
       throw new Error('Question not found')
+    }
+
+    if (authorId !== question.authorId.toValue()) {
+      throw new Error('Not allowed to delete this question')
     }
 
     await this.questionsRepository.delete(question)
